requirements.txt
streamlit
google-generativeai
pypdf
langchain
langchain-community
chromadb
sentence-transformers
python-dotenv